Move typscript interface into api-specification
