+++ packages/api-electron/preload.ts
@@ -3,7 +3,7 @@ import app from './src/preload/app';
 import MessageService from './src/preload/message-service';
 import ScreenSnippet from './src/preload/screen-snippet';
 import Window from './src/preload/window';
-import './src/preload/window-interface';
+import './src/interface';
 
 export {
   app,
