+++ packages/api-electron/src/preload/app.ts
@@ -1,5 +1,7 @@
-const ready = () => Promise.resolve();
+class app implements ssf.App {
+  static ready() {
+    return Promise.resolve();
+  }
+}
 
-export default {
-  ready
-};
+export default app;
