+++ packages/api-electron/src/preload/shims.ts
@@ -1,8 +1,6 @@
 const remote = require('electron').remote;
 const eNotify = remote.require('electron-notify');
 
-interface Window { Notification: any; }
-
 window.Notification = function(title, options) {
   if (!options) {
     options = {};
