+++ packages/api-openfin/preload.ts
@@ -3,7 +3,7 @@ import app from './src/app';
 import MessageService from './src/message-service';
 import ScreenSnippet from './src/screen-snippet';
 import Window from './src/window';
-import './src/window-interface';
+import './src/interface';
 
 export {
   app,
