+++ packages/api-openfin/src/app.ts
@@ -1,7 +1,9 @@
-const ready = () => new Promise((resolve) => {
-  fin.desktop.main(resolve);
-});
-
-export default {
-  ready
+class app implements ssf.App {
+  static ready(){
+    return new Promise((resolve) => {
+      fin.desktop.main(resolve);
+    });
+  }
 };
+
+export default app;
