+++ packages/api-openfin/src/index.ts
@@ -1,0 +1,2 @@
+import './window.interface';
+import './message-service.interface';
